Warnings fixes
Fixed the strings warning in DLL exports by transforming the text field of an AText instance to a pointer. Not sure what the warning was about but it was probably because of an export of a string using a particular C runtime over to another. Now it uses a pointer managed by the instance.

Fixed compiler warnings on linux. Its a wonder things compile at all on Windows...

+++ engine/src/SDL/SDLRenderer.cpp
@@ -177,8 +177,8 @@ void SDLRenderer::Draw(ARenderable* sprite)
 
     if (sdlSprite->GetPositionSystem() == VIEWPORT_RELATIVE)
     {
-        dest.x -= this->RenderViewport->CurrentView.X;
-        dest.y -= this->RenderViewport->CurrentView.Y;
+        dest.x -= (int)this->RenderViewport->CurrentView.X;
+        dest.y -= (int)this->RenderViewport->CurrentView.Y;
     }
 
     tex->RefreshSDLTexture(); // Refresh the texture if needed.
